Move Widget(Type)FormAlter to field_event_dispatcher
Also add all missing hooks as todo in the field_event_dispatcher module
file.
