+++ modules/field_event_dispatcher/src/Event/Field/WidgetFormAlterEvent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\hook_event_dispatcher\Event\Form;
+namespace Drupal\field_event_dispatcher\Event\Field;
 
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\hook_event_dispatcher\Event\EventInterface;
