+++ tests/src/Unit/Form/FormEventTest.php
@@ -4,15 +4,11 @@ namespace Drupal\Tests\hook_event_dispatcher\Unit\Form;
 
 use Drupal;
 use Drupal\Core\DependencyInjection\ContainerBuilder;
-use Drupal\Core\Field\FieldDefinitionInterface;
-use Drupal\Core\Field\FieldItemListInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\hook_event_dispatcher\Event\Form\FormAlterEvent;
-use Drupal\hook_event_dispatcher\Event\Form\WidgetFormAlterEvent;
 use Drupal\hook_event_dispatcher\HookEventDispatcherInterface;
 use Drupal\Tests\hook_event_dispatcher\Unit\HookEventDispatcherManagerSpy;
 use Drupal\Tests\UnitTestCase;
-use function hook_event_dispatcher_field_widget_form_alter;
 use function hook_event_dispatcher_form_alter;
 
 /**
@@ -111,67 +107,6 @@ class FormEventTest extends UnitTestCase {
     $this->assertEquals($form, $event->getForm());
     $this->assertEquals($formState, $event->getFormState());
     $this->assertEquals($formId, $event->getFormId());
-  }
-
-  /**
-   * Test WidgetFormAlterEvent.
-   */
-  public function testWidgetFormAlterEvent(): void {
-    $element = $expectedElement = ['widget' => 'element'];
-    $formState = $this->createMock(FormStateInterface::class);
-    $items = $this->createMock(FieldItemListInterface::class);
-    $definition = $this->createMock(FieldDefinitionInterface::class);
-    $widgetType = 'widget_type';
-    $definition->method('getType')
-      ->willReturn($widgetType);
-    $items->method('getFieldDefinition')
-      ->willReturn($definition);
-    $context = ['items' => $items];
-
-    $this->manager->setEventCallbacks([
-      HookEventDispatcherInterface::WIDGET_FORM_ALTER => static function (WidgetFormAlterEvent $event) {
-        $element = &$event->getElement();
-        $element['other'] = 'key';
-      },
-    ]);
-    $this->manager->setMaxEventCount(2);
-
-    hook_event_dispatcher_field_widget_form_alter($element, $formState, $context);
-
-    $expectedElement['other'] = 'key';
-    $this->assertSame($expectedElement, $element);
-
-    /* @var \Drupal\hook_event_dispatcher\Event\Form\WidgetFormAlterEvent $event */
-    $event = $this->manager->getRegisteredEvent(HookEventDispatcherInterface::WIDGET_FORM_ALTER);
-    $this->assertEquals($element, $event->getElement());
-    $this->assertEquals($formState, $event->getFormState());
-    $this->assertEquals($context, $event->getContext());
-  }
-
-  /**
-   * Test WidgetTypeFormAlterEvent.
-   */
-  public function testWidgetTypeFormAlterEvent(): void {
-    $element = ['widget' => 'element'];
-    $formState = $this->createMock(FormStateInterface::class);
-    $items = $this->createMock(FieldItemListInterface::class);
-    $definition = $this->createMock(FieldDefinitionInterface::class);
-    $widgetType = 'widget_type';
-    $definition->method('getType')
-      ->willReturn($widgetType);
-    $items->method('getFieldDefinition')
-      ->willReturn($definition);
-    $context = ['items' => $items];
-
-    $this->manager->setMaxEventCount(2);
-
-    hook_event_dispatcher_field_widget_form_alter($element, $formState, $context);
-
-    /* @var \Drupal\hook_event_dispatcher\Event\Form\WidgetTypeFormAlterEvent $event */
-    $event = $this->manager->getRegisteredEvent("hook_event_dispatcher.widget_$widgetType.alter");
-    $this->assertEquals($element, $event->getElement());
-    $this->assertEquals($formState, $event->getFormState());
-    $this->assertEquals($context, $event->getContext());
   }
 
 }
